working with new specs

+++ ugine/src/Buffer.h
@@ -7,11 +7,11 @@
 class Buffer {
 	GLuint _vbo = 0;
 	GLuint _ebo = 0;
-	GLuint _vboSize = 0;
-	GLuint _eboSize = 0;
+    size_t _vboSize = 0;
+    size_t _eboSize = 0;
 public:
 
-	Buffer(Vertex* vertexData, GLuint numVertex, GLushort* indexData, GLuint numIndex);
+	Buffer(Vertex* vertexData, size_t numVertex, GLushort* indexData, size_t numIndex);
 	~Buffer();
 	void draw(const Shader& shader) const;
 };
